Externalised load so it can be defined per environment Fixes #19
